+++ AccountManagement.Application.Constracts/Role/EditRole.cs
@@ -2,7 +2,7 @@
 {
    public class EditRole
    {
-      public sbyte Id { get; set; }
+      public int Id { get; set; }
       public string Name { get; set; }
       public string CreationDate { get; set; }
    }
